requirements.txt
Pillow
openai
ultralytics
